+++ api/src/main/java/com/tinqinacademy/bffservice/api/base/OperationProcessor.java
@@ -1,6 +1,6 @@
 package com.tinqinacademy.bffservice.api.base;
 
-import com.tinqinacademy.bffservice.api.error.Errors;
+import com.tinqinacademy.bffservice.api.exceptions.Errors;
 import io.vavr.control.Either;
 
 public interface OperationProcessor<I extends  OperationInput, O extends OperationOutput > {
