+++ api/src/main/java/com/tinqinacademy/bffservice/api/exceptions/Error.java
@@ -1,4 +1,4 @@
-package com.tinqinacademy.bffservice.api.error;
+package com.tinqinacademy.bffservice.api.exceptions;
 
 import lombok.*;
