+++ api/src/main/java/com/tinqinacademy/bffservice/api/exceptions/custom/CustomException.java
@@ -1,4 +1,4 @@
-package com.tinqinacademy.bffservice.core.exceptions.custom;
+package com.tinqinacademy.bffservice.api.exceptions.custom;
 
 import org.springframework.http.HttpStatus;
