+++ core/src/main/java/com/tinqinacademy/bffservice/core/exceptions/ExceptionService.java
@@ -1,7 +1,7 @@
 package com.tinqinacademy.bffservice.core.exceptions;
 
 
-import com.tinqinacademy.bffservice.api.error.Errors;
+import com.tinqinacademy.bffservice.api.exceptions.Errors;
 
 public interface ExceptionService {
     Errors handle(Throwable throwable);
