+++ rest/src/main/java/com/tinqinacademy/bffservice/rest/controllers/BaseController.java
@@ -1,7 +1,7 @@
 package com.tinqinacademy.bffservice.rest.controllers;
 
 import com.tinqinacademy.bffservice.api.base.OperationOutput;
-import com.tinqinacademy.bffservice.api.error.Errors;
+import com.tinqinacademy.bffservice.api.exceptions.Errors;
 import io.vavr.control.Either;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
